connmgr: switch to using net.Addr interface throughout for addresses
This commit modifies the `ConnManager` to use the `net.Add` interface
through the package instead of a plain string to represent and
manipulate addresses. This change makes the package much more general
as users of the package can possibly utilize custom implementations of
the `net.Addr` interface to establish connections.

More precisely, the `ConnReq` struct has been modified to use a
net.Addr instance explicitly, and the `DialFunc` type has also been
modified to take a `net.Addr` directly. This latter change gives
functions that adhere to the `DialFunc` type more flexibility as to
exactly how the connection is established.

In order to main consistency with btcd's IP lookup behavior, a new closure
function for performing DNS queries has been added to to `connmgr.Config`.